+++ cmd/idb.go
@@ -1,3 +1,6 @@
+// Interface for database access, plus migration table creation.
+// AGPL license. Copyright 2024 K Cartlidge.
+
 package main
 
 import "fmt"
@@ -12,8 +15,18 @@ type idb interface {
 	// unless it already exists.
 	ensureMigrationsTable() error
 
+	// removeMigrationsTable drops the `migratable_state` table.
+	removeMigrationsTable() error
+
 	// getCurrentVersion retrieves the version number (0+).
 	getCurrentVersion() (int, error)
+
+	// execMigration executes the migration and updates the
+	// `migratable_state` table accordingly in a transaction.
+	execMigration(statement string, setVersion int, name string, direction string) error
+
+	// exec executes the given SQL statement(s) in a transaction.
+	exec(statement ...string) error
 }
 
 // ensureDB checks it can connect to the database and